MxShop/requirements.txt
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.7
django-cors-headers==3.2.0
django-crispy-forms==1.8.1
django-filter==2.2.0
django-formtools==2.2
django-guardian==2.1.0
django-import-export==2.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
drf-extensions==0.5.0
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.14.0
idna==2.8
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.3
Markdown==3.1.1
MarkupPy==1.14
MarkupSafe==1.1.1
odfpy==1.4.0
openpyxl==3.0.2
Pillow==6.2.1
psycopg2==2.8.4
Pygments==2.5.2
pygments-agentspeak==1.0
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2019.3
PyYAML==5.2
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
tablib==0.14.0
uritemplate==3.0.0
urllib3==1.25.7
xlrd==1.2.0
xlwt==1.3.0
